Lots of changes / improvements
- Notes are dynamically built for all scale types, using the Nashville
Interval as a reference point for each scale note. Currently, this
works correctly for every scale that does not have a double accidental
as its root.

- The Interval class was split into two types: NashvilleInterval and
Interval (for general comparison between two notes).

- Replaced int[] and Interval[] with NashvilleInterval[] in ScaleType.
The type safety of NashvilleInterval guarantees the correct relative
pitch at a given degree (point) in the scale.

- Added a general Theory class to handle processing, and for clearer
organization.

- Added a bunch of scale types. There’s plenty more where that came
from.

- Enabled a few theoretical (i.e. impractical) key signatures.