IntervalSet.setNoteTypes() now throws Exception rather than printing errors and continuing
I haven’t seen any errors for a long time with
IntervalSet.setNoteTypes(), but I want any edge cases to be brought to
my attention with an Exception.

This is because an improperly initialized IntervalSet would cause error
propagation at any location that the faulty IntervalSet is used.

Basically, I consider an error in this algorithm to be serious enough
to warrant an Exception being thrown.