better abstraction
- Meters are subclassed into FixedMeter, TransientMeter, or
InfiniteMeter. This is a top-level differentiator based on what
information the Meter contains with respect to time.

- FixedMeter are subclassed into CompleteMeter and IncompleteMeter.
This determines whether the Meter is counted to a full measure, or to
some partial amount of its beat value.

- CompleteMeters are subclassed into AdditiveCompleteMeter and
NonAdditiveCompleteMeter. This classification decides whether the beat
pattern will appear in the TimeSignature as a summation.

- TransientMeters and IncompleteMeters are not supported yet

- There is only one type of InfiniteMeter: the FreeMeter. It indicates
a meter without a time signature, and thus no bar lines in the staff
for music.

- Logic updated for inferring MeterClassificationTypes

- More MeterClassificationTypes added

- Added Apache Math library for future use with FractionalMeters